requirements.txt
uvicorn[standard]>=0.24.0
fastapi>=0.104.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
requests>=2.31.0
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.13.0
httpx>=0.25.0
aiofiles>=23.0.0
python-multipart>=0.0.6
jinja2>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.0
pytz>=2023.3
